move button tests in to separate spec file that can be used upstream with a custom render function
